Move download handlers outside of react
