+++ app/src/components/GraphContextMenu.tsx
@@ -15,9 +15,14 @@ import { Item, Menu, Separator, Submenu } from "react-contexify";
 import { FiDownload } from "react-icons/fi";
 import { HiOutlineClipboardCopy } from "react-icons/hi";
 
-import { defaultGraphTheme, useCurrentTheme } from "../lib/graphThemes";
+import {
+  defaultGraphTheme,
+  getTheme,
+  useCurrentTheme,
+  useTheme,
+} from "../lib/graphThemes";
 import { borderStyles, shapes } from "../lib/graphUtilityClasses";
-import { useIsFirefox } from "../lib/hooks";
+import { useDownloadFilename, useIsFirefox } from "../lib/hooks";
 import {
   track_downloadJPG,
   track_downloadPng,
@@ -27,6 +32,13 @@ import { useParser } from "../lib/parsers";
 import { useContextMenuState } from "../lib/useContextMenuState";
 import { useDoc } from "../lib/useDoc";
 import { Box, Type } from "../slang";
+import {
+  copyPng,
+  downloadJpg,
+  downloadPng,
+  downloadSvg,
+  getSvg,
+} from "./downloads";
 import styles from "./GraphContextMenu.module.css";
 import { smallIconSize } from "./Shared";
 
@@ -34,6 +46,8 @@ export const GRAPH_CONTEXT_MENU_ID = "graph-context-menu";
 
 export const GraphContextMenu = memo(function GraphContextMenu() {
   const isFirefox = useIsFirefox();
+  const filename = useDownloadFilename();
+  const theme = useTheme();
   return (
     <Menu
       id={GRAPH_CONTEXT_MENU_ID}
@@ -52,7 +66,12 @@ export const GraphContextMenu = memo(function GraphContextMenu() {
       <Separator />
       <Item
         onClick={() => {
-          window.__FF_downloadPNG();
+          if (!theme || !window.__cy) return;
+          downloadPng({
+            filename,
+            theme,
+            cy: window.__cy,
+          });
           track_downloadPng();
         }}
       >
@@ -62,7 +81,12 @@ export const GraphContextMenu = memo(function GraphContextMenu() {
       </Item>
       <Item
         onClick={() => {
-          window.__FF_downloadJPG();
+          if (!theme || !window.__cy) return;
+          downloadJpg({
+            filename,
+            theme,
+            cy: window.__cy,
+          });
           track_downloadJPG();
         }}
       >
@@ -71,8 +95,19 @@ export const GraphContextMenu = memo(function GraphContextMenu() {
         </WithIcon>
       </Item>
       <Item
-        onClick={() => {
-          window.__FF_downloadSVG();
+        onClick={async () => {
+          const theme = getTheme();
+          const cy = window.__cy;
+          if (!theme || !cy) return;
+          const svg = await getSvg({
+            cy,
+            theme,
+          });
+          if (!svg) return;
+          downloadSvg({
+            svg,
+            filename,
+          });
           track_downloadSvg();
         }}
       >
@@ -90,17 +125,24 @@ function CopySVG() {
     (_state: ItemState, action: ItemState) => action,
     "idle"
   );
-  function handleClick() {
-    (async () => {
-      dispatch("loading");
-      const svgStr = await window.__FF_getSVG();
-      // copy to clipboard using navigator
-      await navigator.clipboard.writeText(svgStr);
-      dispatch("success");
-    })();
-  }
+
   return (
-    <Item onClick={handleClick}>
+    <Item
+      onClick={async () => {
+        dispatch("loading");
+        const theme = getTheme();
+        const cy = window.__cy;
+        if (theme && cy)
+          // copy to clipboard using navigator
+          await navigator.clipboard.writeText(
+            await getSvg({
+              cy,
+              theme,
+            })
+          );
+        dispatch("success");
+      }}
+    >
       <WithIcon icon={<HiOutlineClipboardCopy size={smallIconSize} />}>
         <Trans>Copy SVG Code</Trans>
         {state === "loading" ? "..." : ""}
@@ -110,6 +152,8 @@ function CopySVG() {
 }
 
 function CopyPNG() {
+  const theme = useTheme();
+
   const [state, dispatch] = useReducer(
     (_state: ItemState, action: ItemState) => action,
     "idle"
@@ -117,7 +161,11 @@ function CopyPNG() {
   function handleClick() {
     dispatch("loading");
     setTimeout(() => {
-      window.__FF_copyPNG().then(() => {
+      if (!window.__cy || !theme) return;
+      copyPng({
+        cy: window.__cy,
+        theme,
+      }).then(() => {
         dispatch("success");
       });
     }, 0);
